refactor: Move MostViewedFotter files into the components root
