+++ dotcom-rendering/src/web/components/MostViewedFooter.importable.tsx
@@ -1,29 +1,16 @@
-import React, { Suspense } from 'react';
 import { css } from '@emotion/react';
 
 import { text, headline, from, Breakpoint } from '@guardian/source-foundations';
 
-import { initPerf } from '@root/src/web/browser/initPerf';
 import { AdSlot, labelStyles } from '@root/src/web/components/AdSlot';
-import { Lazy } from '@root/src/web/components/Lazy';
 
 import { useAB } from '@guardian/ab-react';
 import { abTestTest } from '@frontend/web/experiments/tests/ab-test-test';
 import { ArticleDesign } from '@guardian/libs';
 import { decidePalette } from '@root/src/web/lib/decidePalette';
-import { Hide } from '../../Hide';
-import { LeftColumn } from '../../LeftColumn';
-
-const MostViewedFooterData = React.lazy(() => {
-	const { start, end } = initPerf('MostViewedFooterData');
-	start();
-	return import(
-		/* webpackChunkName: "MostViewedFooterData" */ './MostViewedFooterData'
-	).then((module) => {
-		end();
-		return { default: module.MostViewedFooterData };
-	});
-});
+import { Hide } from './Hide';
+import { LeftColumn } from './LeftColumn';
+import { MostViewedFooterData } from './MostViewedFooterData';
 
 const stackBelow = (breakpoint: Breakpoint) => css`
 	display: flex;
@@ -130,15 +117,11 @@ export const MostViewedFooter = ({ sectionName, format, ajaxUrl }: Props) => {
 						<Hide when="above" breakpoint="leftCol">
 							<h2 css={headingStyles}>Most popular</h2>
 						</Hide>
-						<Lazy margin={300}>
-							<Suspense fallback={<></>}>
-								<MostViewedFooterData
-									sectionName={sectionName}
-									palette={palette}
-									ajaxUrl={ajaxUrl}
-								/>
-							</Suspense>
-						</Lazy>
+						<MostViewedFooterData
+							sectionName={sectionName}
+							palette={palette}
+							ajaxUrl={ajaxUrl}
+						/>
 					</div>
 					<div
 						css={css`
