+++ dotcom-rendering/src/web/layouts/InteractiveLayout.tsx
@@ -55,7 +55,7 @@ import {
 import { Island } from '../components/Island';
 import { OnwardsLower } from '../components/OnwardsLower.importable';
 import { OnwardsUpper } from '../components/OnwardsUpper.importable';
-import { MostViewedFooter } from '../components/MostViewed/MostViewedFooter/MostViewedFooter';
+import { MostViewedFooter } from '../components/MostViewedFooter.importable';
 
 const InteractiveGrid = ({ children }: { children: React.ReactNode }) => (
 	<div
