+++ dotcom-rendering/src/web/layouts/StandardLayout.tsx
@@ -55,7 +55,7 @@ import { MostViewedRightWrapper } from '../components/MostViewedRightWrapper.imp
 import { GetMatchStats } from '../components/GetMatchStats.importable';
 import { OnwardsLower } from '../components/OnwardsLower.importable';
 import { OnwardsUpper } from '../components/OnwardsUpper.importable';
-import { MostViewedFooter } from '../components/MostViewed/MostViewedFooter/MostViewedFooter';
+import { MostViewedFooter } from '../components/MostViewedFooter.importable';
 
 const StandardGrid = ({
 	children,
